DATA_MINING_STUDENT_PROJECT/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
mlxtend
pyyaml
joblib
jupyter